+++ convex/auth.ts
@@ -1,3 +1,6 @@
+import Facebook from "@auth/core/providers/facebook";
+import GitHub from "@auth/core/providers/github";
+import Google from "@auth/core/providers/google";
 import { Password } from "@convex-dev/auth/providers/Password";
 import { convexAuth } from "@convex-dev/auth/server";
 
@@ -39,6 +42,9 @@ export const { auth, signIn, signOut, store, isAuthenticated } = convexAuth({
         return out;
       },
     }),
+    GitHub,
+    Google,
+    Facebook,
   ],
   callbacks: {
     async afterUserCreatedOrUpdated(ctx, { userId, existingUserId, profile }) {